+++ src/components/tabs/tabs.js
@@ -2,6 +2,13 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 
+const displayToClassNameMap = {
+  condensed: 'euiTabs--condensed',
+  default: null,
+};
+
+export const DISPLAYS = Object.keys(displayToClassNameMap);
+
 const sizeToClassNameMap = {
   s: 'euiTabs--small',
   m: null,
@@ -10,14 +17,16 @@ const sizeToClassNameMap = {
 export const SIZES = Object.keys(sizeToClassNameMap);
 
 export const EuiTabs = ({
-  size,
-  expand,
   children,
   className,
+  display,
+  expand,
+  size,
   ...rest
 }) => {
   const classes = classNames(
     'euiTabs',
+    displayToClassNameMap[display],
     sizeToClassNameMap[size],
     {
       'euiTabs--expand': expand,
@@ -26,11 +35,7 @@ export const EuiTabs = ({
   );
 
   return (
-    <div
-      role="tablist"
-      className={classes}
-      {...rest}
-    >
+    <div role="tablist" className={classes} {...rest}>
       {children}
     </div>
   );
@@ -39,15 +44,20 @@ export const EuiTabs = ({
 EuiTabs.propTypes = {
   children: PropTypes.node,
   className: PropTypes.string,
-  size: PropTypes.oneOf(SIZES),
+  /**
+   * Choose `default` or alternative `condensed` display styles
+   */
+  display: PropTypes.oneOf(DISPLAYS),
   /**
    * Evenly stretches each tab to fill the
    * horizontal space
    */
   expand: PropTypes.bool,
+  size: PropTypes.oneOf(SIZES),
 };
 
 EuiTabs.defaultProps = {
-  size: 'm',
+  display: 'default',
   expand: false,
+  size: 'm',
 };